// lib/schemas.ts
export const cardSchema = z.object({
  transactionId: z.string(),
  memberId: z.string(),
  status: z.enum(["PENDING", "SUCCEEDED", "FAILED"]),
  address: z.string(),
  address2: z.string().optional(),
  city: z.string(),
  region: z.string(),
  postalCode: z.string(),
  countryCode: z.string(),
  name: z.string(),
  phoneNumber: z.string(),
  email: z.string(),

  cardFrontPath: z.string().optional(),
  cardBackPath: z.string().optional(),
  printagsCardFrontId: z.string().optional(),
  printagsCardBackId: z.string().optional(),
  adminKey: z.string(),
  counter: z.number().default(0),
  privateKey: z.string(),

  id: z.string(),
  createdAt: z.date(),
  updatedAt: z.date(),
  deletedAt: z.date().optional(),
  member: z.any().nullish(),
});
// ...
export const memberSchema = z.object({
  firstName: z.string().min(2, {
    message: "Le prénom doit faire au moins 2 caractères",
  }),
  lastName: z.string().optional(),
  phoneNumber: z.string().optional(),
  membershipExpiresAt: z.date().nullish(),
  isMembershipActive: z.boolean().default(true),
  email: z.string().optional(),
  imagePath: z.string().optional(),
  note: z.string().default(""),
  id: z.string(),
  organizationId: z.string(),
  position: z.string().optional(),
  birthDate: z.date().optional(),
  nationality: z.string().optional(),

  address: z.string().optional(),
  address2: z.string().optional(),
  city: z.string().optional(),
  region: z.string().optional(),
  postalCode: z.string().optional(),
  countryCode: z.string().optional(),

  createdAt: z.date(),
  updatedAt: z.date(),
  deletedAt: z.date().optional(),
  cards: z.array(cardSchema).optional(),
});

export type Member = z.infer<typeof memberSchema>;